+++ docs/ReleaseNotes.md
@@ -7,6 +7,7 @@
 
 ### Version 1.1.22
 + Light: Added redThreshold, greenThreshold, blueThreshold for when whiteMix is false in an RGBWW light
++ Light: Added minColorTemperature and maxColorTemperature configuration settings
 
 ### Version 1.1.21
 + Updated fakegato-history dependency version